requirements-api.txt
fastapi
uvicorn
langchain
langchain-community
chromadb
gitpython
python-dotenv
tqdm
